Ajustes Nas lógicas do player e saneamento de códigos... foram removidos alguns inimigos para facilitar os testes ainda existem problemas com as escadas (mas são poucos)

+++ src/entidades/naoSolidos/MountainsParalax.java
@@ -16,7 +16,7 @@ public class MountainsParalax extends Entity {
         // velocidade de deslocamento da montanha
 
         if (Camera.x - (Player.atualX / 8) > 0) {
-            this.x = (int) (Camera.x - (Player.atualX / 8));
+            this.x = Camera.x - (Player.atualX / 8);
         } else {
             this.x = 0;
         }
